Adds Fritz!Box API cache option
Enables the use of API caching for the Fritz!Box connection.

This change introduces a `use_cache` option in the configuration,
allowing users to enable or disable API caching to improve performance.

The new option is added to the configuration file, model, and application,
and passed to the FritzConnection library.